refactor: extract test data seeding into reusable helper function
