+++ apps/client/src/__tests__/counts-mongo-vs-legacy.test.ts
@@ -2,6 +2,7 @@ import { MongoMemoryServer } from "mongodb-memory-server";
 import mongoose, { Connection, Schema } from "mongoose";
 import { computeSearchCounts, QueryParams } from "~/pages/api/search/counts";
 import { legacyFacetCounts, LegacyNeedsItem, LegacyQuery } from "~/tests/legacyCounts";
+import { seedDispositifs } from "~/tests/seedDispositifs";
 
 // Mock Algolia client to reflect @algolia/client-search usage and avoid real network
 jest.mock("@algolia/client-search", () => {
@@ -65,46 +66,7 @@ describe("Mongo counts vs legacy filterDispositifs", () => {
     const db = conn.db;
     if (!db) throw new Error("Connection DB not initialized");
     await db.dropDatabase();
-
-    const Dispositif = conn.model("Dispositif");
-    await Dispositif.insertMany([
-      // Paris (75), FR/EN, public: [jeunes], french A1, themeA, needs [A1,A2], age 16-25
-      {
-        thematiques: [themeA],
-        besoins: [needA1, needA2],
-        metadatas: { location: "75", frenchLevel: ["A1"], public: ["jeunes"], age: { from: 16, to: 25 } },
-        availableLanguages: ["fr", "en"],
-        status: "Actif",
-        typeContenu: "dispositif",
-      },
-      // Hauts-de-Seine (92), FR only, public: [familles], french B1, themeA, needs [A1], age 26-64
-      {
-        thematiques: [themeA],
-        besoins: [needA1],
-        metadatas: { location: "92", frenchLevel: ["B1"], public: ["familles"], age: { from: 26, to: 64 } },
-        availableLanguages: ["fr"],
-        status: "Actif",
-        typeContenu: "dispositif",
-      },
-      // Paris (75), FR, public: [seniors], french A2, themeB, needs [B1], age 65+
-      {
-        thematiques: [themeB],
-        besoins: [needB1],
-        metadatas: { location: "75", frenchLevel: ["A2"], public: ["seniors"], age: { from: 65, to: 90 } },
-        availableLanguages: ["fr"],
-        status: "Actif",
-        typeContenu: "dispositif",
-      },
-      // Inactive entry should be filtered out globally
-      {
-        thematiques: [themeB],
-        besoins: [needB1],
-        metadatas: { location: "75", frenchLevel: ["A1"], public: ["adultes"], age: { from: 18, to: 25 } },
-        availableLanguages: ["fr"],
-        status: "Inactif",
-        typeContenu: "dispositif",
-      },
-    ]);
+    await seedDispositifs(conn, { themeA, themeB, needA1, needA2, needB1 });
   });
 
   const toLegacyQuery = (p: Partial<LegacyQuery>): LegacyQuery => ({
